chore(playList/service) :能够播放音乐 因为不能回调点击item的位置 所以导致不能正确播放音乐

+++ utils/src/main/java/com/example/cloudmusic/utils/webs/ServiceCreator.kt
@@ -8,7 +8,7 @@ import java.util.concurrent.TimeUnit
 
 object ServiceCreator {
 
-    private const val TIME_OUT = 30.toLong()
+    private const val TIME_OUT = 30L
 
     private val client = OkHttpClient.Builder()
         .readTimeout(TIME_OUT, TimeUnit.SECONDS)
